refactor(tests): split plane-instance tests into sibling file (#679 CI)
scripts/check-rs-line-limit.sh enforces a 500-line cap on .rs files. The
new plane-instance tests pushed renderer3d/tests.rs to 615 lines, which
made CI's `Rust File Line Limit` job fail. Moved the 5 plane-instance
tests into a sibling `tests_plane_instances.rs` so both files stay under
the cap. No test logic changed.
